separated non-pgsql test
